planner-mcp/pyproject.toml
[project]
name = "planner-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gitignore-parser>=0.1.12",
    "mcp>=1.6.0",
    "python-magic>=0.4.27",
    "tiktoken>=0.9.0",
]
